Add basic validation to UserServiceImpl

+++ src/main/java/pl/umcs/coffee/user/UserServiceImpl.java
@@ -1,6 +1,9 @@
 package pl.umcs.coffee.user;
 
+import org.jetbrains.annotations.NotNull;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -11,7 +14,13 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User createUser(User user) {
+    public User createUser(@NotNull User user) {
+        User foundUser = userRepository.findById(user.getId()).orElse(null);
+
+        if (foundUser != null) {
+            throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY);
+        }
+
         return userRepository.save(user);
     }
 
@@ -21,13 +30,23 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User updateUser(User user) {
+    public User updateUser(@NotNull User user) {
+        User foundUser = userRepository.findById(user.getId()).orElse(null);
+
+        if (foundUser != null) {
+            throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY);
+        }
+
         return userRepository.save(user);
     }
 
     @Override
     public User deleteUser(Long id) {
         User deletedUser = userRepository.findById(id).orElse(null);
+
+        if (deletedUser == null) {
+            throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY);
+        }
         userRepository.deleteById(id);
 
         return deletedUser;
